+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Added error handling to search API
 - Refactored analytics report code
 - Implemented country-level analytics report
+- Implemented FRBRization v2 in OCLC Classify Process
 
 ## Fixed
 - Changed HATHI_DATAFILES outdated link in development, example, and local yaml files